Q3

+++ 0x0B-malloc_free/3-alloc_grid.c
@@ -8,7 +8,7 @@
  */
 int **alloc_grid(int width, int height)
 {
-	int **s;
+	long int **s;
 	int w, h;
 
 	if (width <= 0 || height <= 0)
